+++ CHANGELOG.md
@@ -10,6 +10,7 @@ en-US:
 - Add emphasized color border to the activated session in split screen mode
 - Add custom color feature to the highlight
 - Allow editing the text to be pasted when confirming multiple lines of paste
+- Allow the tab title to display a custom session name
 - Fix the issue that clicking the new tab button in split screen mode may not create the session correctly or be located under the wrong tab group
 - Fix the issue that the SSH connection cannot be reconnected by tapping the Enter key in some cases
 - Fix the issue that the target session object is inaccurate when locking/unlocking the session
@@ -22,6 +23,7 @@ zh-CN:
 - 分屏模式下激活的会话增加强调色边框
 - 高亮功能增加自定义颜色功能
 - 多行粘贴确认时允许编辑待粘贴文本
+- 标签栏标题允许显示自定义会话名称
 - 修复分屏模式下某些情况点击新标签按钮会话未正确创建或位于错误的标签页组下
 - 修复ssh连接部分情况下无法通过敲击回车键发起重连的问题
 - 修复锁定/解锁会话时目标会话对象不准确